# Simplify reply sent monitoring