@jsumners 2021-05-14
Tests fail due to decreased coverage -- https://github.com/fastify/fastify/pull/3072/checks?check_run_id=2585550350#step:6:11141