@jsumners 2021-05-14
> > Tests fail due to decreased coverage -- [#3072 (checks)](https://github.com/fastify/fastify/pull/3072/checks?check_run_id=2585550350#step:6:11141)
> 
> Do you have any suggestion for solving this? I believe it's because of
> 
> [sergiz/fastify@`8e61a47`/lib/reply.js#L85](https://github.com/sergiz/fastify/blob/8e61a472e6d091fab328e425d23ecbb663ed6b2e/lib/reply.js#L85)

Have you run the HTML coverage report to see what is not being covered?