@sergiz 2021-05-14
> Have you run the HTML coverage report to see what is not being covered?

I installed the node v10 and the console already showed me that that line was problematic. I included two more tests now which cover both possibilities.


> Can you do a benchmark compare?

How should I do that? I ran `npm run benchmark` a few times and the results are in the same range which makes sense since the only change that this PR does is to read from a different property.