@RafaelGSS 2021-05-15
> How should I do that? I ran `npm run benchmark` a few times and the results are in the same range which makes sense since the only change that this PR does is to read from a different property.

Yes, `npm run bench` and compare `main` with your branch. Can you share the results?