@sergiz 2021-05-15
> Yes, `npm run bench` and compare `main` with your branch. Can you share the results?

`main`
```
[1] Running 5s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1] 
[1] ┌─────────┬───────┬───────┬───────┬───────┬──────────┬──────────┬────────┐
[1] │ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev    │ Max    │
[1] ├─────────┼───────┼───────┼───────┼───────┼──────────┼──────────┼────────┤
[1] │ Latency │ 11 ms │ 24 ms │ 42 ms │ 73 ms │ 22.12 ms │ 11.36 ms │ 195 ms │
[1] └─────────┴───────┴───────┴───────┴───────┴──────────┴──────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
[1] │ Req/Sec   │ 31631   │ 31631   │ 47295   │ 49151   │ 44257.6 │ 6511.52 │ 31617   │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
[1] │ Bytes/Sec │ 5.91 MB │ 5.91 MB │ 8.85 MB │ 9.19 MB │ 8.28 MB │ 1.22 MB │ 5.91 MB │
[1] └───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘
[1] 
[1] Req/Bytes counts sampled once per second.
[1] 
[1] 221k requests in 5.04s, 41.4 MB read
[1] npx autocannon -c 100 -d 5 -p 10 localhost:3000/ exited with code 0
```

```
[1] Running 5s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1] 
[1] ┌─────────┬──────┬───────┬───────┬───────┬──────────┬──────────┬────────┐
[1] │ Stat    │ 2.5% │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev    │ Max    │
[1] ├─────────┼──────┼───────┼───────┼───────┼──────────┼──────────┼────────┤
[1] │ Latency │ 9 ms │ 21 ms │ 41 ms │ 66 ms │ 20.64 ms │ 10.53 ms │ 123 ms │
[1] └─────────┴──────┴───────┴───────┴───────┴──────────┴──────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬────────┬─────────┬─────────┬─────────┬────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%    │ 97.5%   │ Avg     │ Stdev   │ Min    │
[1] ├───────────┼─────────┼─────────┼────────┼─────────┼─────────┼─────────┼────────┤
[1] │ Req/Sec   │ 35839   │ 35839   │ 48671  │ 54815   │ 47344   │ 6295.15 │ 35838  │
[1] ├───────────┼─────────┼─────────┼────────┼─────────┼─────────┼─────────┼────────┤
[1] │ Bytes/Sec │ 6.71 MB │ 6.71 MB │ 9.1 MB │ 10.3 MB │ 8.85 MB │ 1.18 MB │ 6.7 MB │
[1] └───────────┴─────────┴─────────┴────────┴─────────┴─────────┴─────────┴────────┘
[1] 
[1] Req/Bytes counts sampled once per second.
[1] 
[1] 237k requests in 5.04s, 44.3 MB read
[1] npx autocannon -c 100 -d 5 -p 10 localhost:3000/ exited with code 0
```

`pr/simplify-reply-sent-monitoring`
```
[1] Running 5s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1] 
[1] ┌─────────┬───────┬───────┬───────┬───────┬──────────┬──────────┬────────┐
[1] │ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev    │ Max    │
[1] ├─────────┼───────┼───────┼───────┼───────┼──────────┼──────────┼────────┤
[1] │ Latency │ 10 ms │ 21 ms │ 45 ms │ 67 ms │ 21.65 ms │ 11.53 ms │ 188 ms │
[1] └─────────┴───────┴───────┴───────┴───────┴──────────┴──────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
[1] │ Req/Sec   │ 33855   │ 33855   │ 47455   │ 52127   │ 45206.4 │ 6339.41 │ 33843   │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
[1] │ Bytes/Sec │ 6.33 MB │ 6.33 MB │ 8.88 MB │ 9.75 MB │ 8.45 MB │ 1.19 MB │ 6.33 MB │
[1] └───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘
[1] 
[1] Req/Bytes counts sampled once per second.
[1] 
[1] 226k requests in 5.04s, 42.3 MB read
[1] npx autocannon -c 100 -d 5 -p 10 localhost:3000/ exited with code 0
```

```
[1] Running 5s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1] 
[1] ┌─────────┬──────┬───────┬───────┬───────┬──────────┬──────────┬────────┐
[1] │ Stat    │ 2.5% │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev    │ Max    │
[1] ├─────────┼──────┼───────┼───────┼───────┼──────────┼──────────┼────────┤
[1] │ Latency │ 9 ms │ 20 ms │ 39 ms │ 63 ms │ 20.41 ms │ 10.77 ms │ 150 ms │
[1] └─────────┴──────┴───────┴───────┴───────┴──────────┴──────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
[1] │ Req/Sec   │ 35359   │ 35359   │ 50687   │ 54655   │ 47862.4 │ 6787.31 │ 35351   │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
[1] │ Bytes/Sec │ 6.61 MB │ 6.61 MB │ 9.48 MB │ 10.2 MB │ 8.95 MB │ 1.27 MB │ 6.61 MB │
[1] └───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘
[1] 
[1] Req/Bytes counts sampled once per second.
[1] 
[1] 239k requests in 5.04s, 44.7 MB read
[1] npx autocannon -c 100 -d 5 -p 10 localhost:3000/ exited with code 0
```