@mcollina 2021-05-29
Unfortunately this is a breaking change, I'll revert on main and reapply it on `next`.